+++ hyperbench/tests/utils/nn_utils_test.py
@@ -43,11 +43,12 @@ def test_maxmin_scatter_computes_channelwise_range_by_group():
             [-2.0, 7.0],
             [5.0, -1.0],
             [5.0, 8.0],
-        ]
+        ],
+        dtype=torch.float,
     )
     # index[k] says which output group receives src[k].
     # Example: index[1] == 0 means src[1] = [3, 1] contributes to output row 0.
-    index = torch.tensor([0, 0, 1, 1, 1])
+    index = torch.tensor([0, 0, 1, 1, 1], dtype=torch.long)
 
     # dim=0 scatters rows into grouped output rows, preserving the feature/channel dimension.
     result = maxmin_scatter(src=src, index=index, dim=0)
@@ -60,7 +61,8 @@ def test_maxmin_scatter_computes_channelwise_range_by_group():
         [
             [2.0, 3.0],
             [7.0, 9.0],
-        ]
+        ],
+        dtype=torch.float,
     )
     assert torch.allclose(result, expected)
 
@@ -71,7 +73,8 @@ def test_maxmin_scatter_respects_explicit_dim_size():
             [1.0, 4.0],
             [3.0, 1.0],
             [-2.0, 7.0],
-        ]
+        ],
+        dtype=torch.float,
     )
 
     # index[k] says which output group receives src[k].
@@ -79,7 +82,7 @@ def test_maxmin_scatter_respects_explicit_dim_size():
     # - index[1] == 0 means src[1] = [3, 1] contributes to output row 0.
     # - index[2] == 2 means src[2] = [-2, 7] contributes to output row 2.
     # Missing group ids indicate that those groups receive no source rows, so group 1 and group 3 are empty.
-    index = torch.tensor([0, 0, 2])
+    index = torch.tensor([0, 0, 2], dtype=torch.long)
 
     # dim_size=4 forces four output rows even though max(index) would only imply three rows.
     result = maxmin_scatter(src=src, index=index, dim=0, dim_size=4)
@@ -94,7 +97,8 @@ def test_maxmin_scatter_respects_explicit_dim_size():
             [0.0, 0.0],
             [0.0, 0.0],
             [0.0, 0.0],
-        ]
+        ],
+        dtype=torch.float,
     )
 
     assert torch.allclose(result, expected)
@@ -105,13 +109,14 @@ def test_maxmin_scatter_supports_nonzero_scatter_dimension():
         [
             [1.0, 5.0, 3.0],
             [4.0, 2.0, 8.0],
-        ]
+        ],
+        dtype=torch.float,
     )
 
     # With dim=1, index[k] says which output column group receives source column k.
     # Example: index[2] == 0 means the third source column contributes to output column 0.
     # Source columns 0 and 2 are grouped together, while source column 1 is alone in group 1.
-    index = torch.tensor([0, 1, 0])
+    index = torch.tensor([0, 1, 0], dtype=torch.long)
 
     # dim_size=2 keeps exactly two output column groups: group 0 and group 1.
     result = maxmin_scatter(src=src, index=index, dim=1, dim_size=2)
@@ -124,7 +129,8 @@ def test_maxmin_scatter_supports_nonzero_scatter_dimension():
         [
             [2.0, 0.0],
             [4.0, 0.0],
-        ]
+        ],
+        dtype=torch.float,
     )
     assert torch.allclose(result, expected)
